+++ .storybook/main.ts
@@ -11,15 +11,14 @@ const config: StorybookConfig = {
     "../src/docs/colors/colors.mdx",
     "../src/docs/**/*.stories.@(js|jsx|mjs|ts|tsx)",
     "../tedi/**/*.stories.@(js|jsx|mjs|ts|tsx)",
+    "../tedi/**/*.mdx",
     "../community/**/*.stories.@(js|jsx|mjs|ts|tsx)",
   ],
   addons: [
-    "@storybook/addon-essentials",
-    "@storybook/addon-interactions",
+    "@storybook/addon-docs",
     "@storybook/addon-a11y",
     "@etchteam/storybook-addon-status",
     "storybook-addon-pseudo-states",
-    "storybook-addon-angular-router",
   ],
   framework: {
     name: "@storybook/angular",
@@ -28,9 +27,6 @@ const config: StorybookConfig = {
     },
   },
   staticDirs: ["../public"],
-  docs: {
-    autodocs: true,
-  },
   core: {
     disableTelemetry: true,
   },