+++ tedi/components/buttons/index.ts
@@ -3,5 +3,6 @@ export * from "./button-group/button-group.component";
 export * from "./button-group/button-group-item/button-group-item.directive";
 export * from "./info-button/info-button.component";
 export * from "./collapse/collapse.component";
+export * from "./collapse-button/collapse-button.component";
 export * from "./closing-button/closing-button.component";
 export * from "./button/base-button.directive";